fix(evaluations): compare Metadata by value so a reformat isn't a change
